third_party/blink/web_tests/platform/generic/svg/dynamic-updates/SVGFEComponentTransferElement-svgdom-slope-prop-expected.txt

SVG 1.1 dynamic update tests

Tests dynamic updates of the 'slope' property of the SVGFEComponentTransferElement object

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS feRFunc.slope.baseVal is 1
PASS feGFunc.slope.baseVal is 1
PASS feBFunc.slope.baseVal is 1
PASS feAFunc.slope.baseVal is 1
PASS feRFunc.slope.baseVal is 2
PASS feGFunc.slope.baseVal is 2
PASS feBFunc.slope.baseVal is 2
PASS feAFunc.slope.baseVal is 2
PASS successfullyParsed is true

TEST COMPLETE
